ml_system/scripts/requirements_ml.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
python-dotenv>=0.19.0